requirements.txt
fastapi==0.110.0
streamlit==1.32.2
python-dotenv==1.0.1
uvicorn==0.27.1
python-multipart==0.0.9
pillow==10.2.0
torch==2.2.1
torchvision==0.17.1
numpy==1.26.4
scikit-learn==1.4.1.post1 